refactor: 헤더 및 TabBar 그룹화 (#66)
