+++ apps/web/src/app/(with-global-layout)/auctions/write/page.tsx
