+++ apps/web/src/app/(with-global-layout)/auth/layout.tsx
