+++ apps/web/src/app/(with-local-layout)/(with-header-layout)/layout.tsx
@@ -1,0 +1,13 @@
+import React, { ReactNode } from 'react';
+import Header from 'src/components/layout/Header';
+
+const HeaderLayout = ({ children }: { children: ReactNode }) => {
+  return (
+    <>
+      <Header />
+      <div>{children}</div>
+    </>
+  );
+};
+
+export default HeaderLayout;
