+++ apps/web/src/app/(with-local-layout)/(with-header-layout)/mypage/page.tsx
