+++ apps/web/src/app/(with-local-layout)/auctions/[id]/page.tsx
