+++ apps/web/src/app/(with-local-layout)/mypage/points/page.tsx
