+++ apps/web/src/app/layout.tsx
@@ -1,10 +1,8 @@
-import QueryProvider from '../providers/QueryProvider';
-import { home } from '../metadata/common/home';
-import './globals.css';
-import TabBar from '../components/layout/TabBar';
-import Header from 'src/components/layout/Header';
 import { pretendard } from 'assets/fonts';
 import AppProvider from 'src/providers/AppProvider';
+import { home } from '../metadata/common/home';
+import QueryProvider from '../providers/QueryProvider';
+import './globals.css';
 
 export const metadata = home;
 
@@ -14,9 +12,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       <body className={`${pretendard.className}`}>
         <div className="relative m-auto flex w-full max-w-2xl flex-col shadow-2xl">
           <QueryProvider>
-            <Header />
             {children}
-            <TabBar />
             <AppProvider />
           </QueryProvider>
         </div>
